fix: undefined unit list breaks js in client
Problem:
If an agent has an empty list of units it will be undefined when we GET
list of units for an agent group. This will cause a js error when we try
to forEach the undefined array (unit-collectinon.js row 86).

As a result the list of deployable units will not be updated in the UI when
the user chooses such a group in the top right corner of the dashboard.

Solution:
Assign an empty array to the units property if agent has no units.